Before cancelling account, check if he is admin with members. Only admin can cancel accounts. Link to cancel account removed from edit_profile.

+++ app/controllers/profiles_controller.rb
@@ -65,6 +65,10 @@ class ProfilesController < ApplicationController
       redirect_to :back, alert: 'Can not delete own user.'
       return
     end
+    if(usr.members.count > 0)
+      redirect_to :back, alert: 'Can not delete admin user as there are allocated members.'
+      return
+    end
     name = usr.name
     usr.destroy
     redirect_to :back, notice: 'User '+name+' removed'
